Do Firmware Upgrade only on Raspberrys

+++ sbin/loxberryupdate/updatereboot_v3.0.0.pl
@@ -248,27 +248,29 @@ if ( -e "/etc/logrotate.conf.dpkg-dist" ) {
 my $output = qx { sed -i --follow-symlinks 's/^#compress/compress/g' /etc/logrotate.conf >> $logfilename 2>&1 };
 $log->open;
 
-# Update Kernel and Firmware
+# Update Kernel and Firmware on Raspberry
 # GIT Firmware Hash:   d5edc6af1ef48f97b525da88ff6c510c2d4231c3
 # dirtree Checksum is: 7d32024a09eac34dfe5e511ebbc01af7
 rpi_update("7d32024a09eac34dfe5e511ebbc01af7", "d5edc6af1ef48f97b525da88ff6c510c2d4231c3");
 
 #
-# Firmware Files are not updated automatically by apt-get (why? *really* don't no!)
+# Firmware Files are not updated automatically by apt-get (why? *really* don't know!) - only on Raspberry
 #
-LOGINF "Installing newest firmware files (from Debian Buster because Bullseye isn't currently available)...";
-# Use RPi-Distro Repo
-system("curl -L https://github.com/RPi-Distro/firmware-nonfree/archive/refs/heads/buster.zip -o /lib/master.zip");
-system("cd /lib && unzip /lib/master.zip");
-$exitcode  = $? >> 8;
-if ($exitcode != 0) {
+if (-e "$LoxBerry::System::lbhomedir/config/system/is_raspberry.cfg" && !-e "$LoxBerry::System::lbhomedir/config/system/is_odroidxu3xu4.cfg") {
+	LOGINF "Installing newest firmware files (from Debian Buster because Bullseye isn't currently available)...";
+	# Use RPi-Distro Repo
+	system("curl -L https://github.com/RPi-Distro/firmware-nonfree/archive/refs/heads/buster.zip -o /lib/master.zip");
+	system("cd /lib && unzip /lib/master.zip");
+	$exitcode  = $? >> 8;
+	if ($exitcode != 0) {
         LOGERR "Error extracting new firmware. This is a problem for Zero2 only. Wifi may not work on the Zero2 - Error $exitcode";
-} else {
+	} else {
         LOGOK "Extracting of new firmware files successfully. Installing...";
-	system ("rm -r /lib/firmware");
-	system("mv /lib/firmware-nonfree-buster /lib/firmware");
+		system ("rm -r /lib/firmware");
+		system("mv /lib/firmware-nonfree-buster /lib/firmware");
+	}
+	system ("rm -r /lib/master.zip");
 }
-system ("rm -r /lib/master.zip");
 
 #
 # Reinstall Python packages, because rasbian's upgrade will overwrite all of them...
@@ -331,8 +333,7 @@ qx { chmod +x $lbhomedir/system/daemons/system/98-updaterebootcontinue };
 #
 # Installing new dependencies
 #
-
-LOGINF "Installing new packages";
+LOGINF "Installing new packages for LoxBerry 3.0...";
 apt_install("libcgi-simple-perl");
 
 #
@@ -340,7 +341,10 @@ apt_install("libcgi-simple-perl");
 #
 
 #LOGINF "MQTT Gateway migration";
-# LoxBerry::System::execute( command => '$lbhomedir/sbin/loxberryupdate/mqtt_migration.pl', log => $log );
+#LoxBerry::System::execute( command => '$lbhomedir/sbin/loxberryupdate/mqtt_migration.pl', log => $log );
+
+
+
 
 
 LOGOK "Update script $0 finished." if ($errors == 0);
